V_citations/requirements.txt
openpyxl
pandas
numpy
python-docx
Flask